sentinel_social/backend/requirements.txt
fastapi
uvicorn
torch
git+https://github.com/openai/CLIP.git
requests
python-multipart
pillow
numpy
nltk
plotly
pydantic
